Add regsiter and get_regsitered routes
The register route with post method to
register new modules for cluster. You just
need to give a unique name and manager
will create a unique 64 character secret
key for your module. Then you should
put this secret key in module to use it
for logging in the module to manager
and joining it to cluster.